Move bugzilla service to new jbi.bugzilla.service module
