+++ jbi/bugzilla/__init__.py
@@ -1,0 +1,3 @@
+from .service import BugzillaService, get_service
+
+__all__ = ["get_service", "BugzillaService"]
